requirements.txt
fastapi==0.114.0
uvicorn==0.30.5
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
httpx==0.27.0
python-dotenv==1.0.1
tenacity==8.5.0
structlog==24.1.0
redis==5.0.8
pytest==7.4.0
pytest-asyncio==1.1.0
PyYAML==6.0.1
